Adding serialization and docs
