+++ Cargo.toml
@@ -33,6 +33,7 @@ p256_ = { package = "p256", version = "0.9", default-features = false, features 
 rand = { version = "0.8", default-features = false }
 subtle = { version = "2.3", default-features = false }
 zeroize = { version = "1", features = ["zeroize_derive"] }
+serde = { version = "1", default-features = false, features = ["alloc", "derive"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2", features = ["js"], optional = true }
